Prime Factorization
Prime factorisation or complete factorisation of the given number is to express a given number as a product of prime factor.
When a number is expressed as the product of its prime factors, it is called prime factorization.
For example, 15 = 3 × 5. So, 3 and 5 are prime factors of 15.
Prime factors of a number are always the prime numbers, when a number is expressed as product of prime numbers, it is called prime factorization.

@ Method :
Divide the given numbers by the smallest prime numbers. Continue the division until it is not further divisible of prime numbers.